spark-etl/requirements.txt
pyspark==3.5.0
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
mysql-connector-python==8.2.0
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.0
werkzeug==3.0.1
requests==2.31.0

